Replace old admin app with new MUI admin app (#136)
* temp - remove old `src`

* temp - replace/remove files

* move logo png to inside admin app

* Add condition for catching missing content_id
