+++ admin_app/next.config.mjs
