+++ admin_app/src/app/content/layout.tsx
