--- admin_app/src/components/Config.tsx
@@ -1,2 +1,0 @@
-export const backendUrl: string =
-  process.env.NEXT_PUBLIC_BACKEND_URL || "http://localhost:8000";
