+++ admin_app/src/components/ContentModal.tsx
