+++ admin_app/src/components/ProtectedComponent.tsx
