--- admin_app_2/package.json
@@ -1,31 +1,0 @@
-{
-  "name": "admin-app",
-  "version": "0.2.0",
-  "private": true,
-  "scripts": {
-    "dev": "next dev",
-    "build": "next build",
-    "start": "next start",
-    "lint": "next lint"
-  },
-  "dependencies": {
-    "@emotion/react": "^11.11.3",
-    "@emotion/styled": "^11.11.0",
-    "@fontsource/roboto": "^5.0.12",
-    "@mui/icons-material": "^5.15.10",
-    "@mui/material": "^5.15.10",
-    "jwt-decode": "^4.0.0",
-    "next": "14.1.3",
-    "react": "^18",
-    "react-dom": "^18"
-  },
-  "devDependencies": {
-    "@types/node": "^20",
-    "@types/react": "^18",
-    "@types/react-dom": "^18",
-    "eslint": "^8",
-    "eslint-config-next": "14.1.3",
-    "prettier-plugin-organize-imports": "^3.2.4",
-    "typescript": "^5"
-  }
-}
