--- admin_app_2/src/app/layout.tsx
@@ -1,32 +1,0 @@
-import theme from "@/theme";
-import AuthProvider from "@/utils/auth";
-import { CssBaseline, ThemeProvider } from "@mui/material";
-import type { Metadata } from "next";
-import { Inter } from "next/font/google";
-import React from "react";
-const inter = Inter({ subsets: ["latin"] });
-
-export const metadata: Metadata = {
-  title: "Ask A Question",
-  description: "Admin application to manage content and glean insights",
-};
-
-export default function RootLayout({
-  children,
-}: Readonly<{
-  children: React.ReactNode;
-}>) {
-  return (
-    <html lang="en">
-      <head>
-        <meta name="viewport" content="initial-scale=1, width=device-width" />
-      </head>
-      <body className={inter.className}>
-        <ThemeProvider theme={theme}>
-          <CssBaseline />
-          <AuthProvider>{children}</AuthProvider>
-        </ThemeProvider>
-      </body>
-    </html>
-  );
-}
