--- admin_app_2/src/app/page.tsx
@@ -1,14 +1,0 @@
-"use client";
-
-import { useRouter } from "next/navigation";
-import { useEffect } from "react";
-
-export default function Home() {
-  const router = useRouter();
-
-  useEffect(() => {
-    router.push("/content");
-  }, []);
-
-  return <div></div>;
-}
